Apply improved types & architecture to ICP
